% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejSamp.R
\name{rejSamp}
\alias{rejSamp}
\title{Generate random numbers by rejection sampling}
\usage{
rejSamp(f, n = 1, min = 0, max = 1, g = NULL, rg = NULL,
  g.factor = 1)
}
\arguments{
\item{f}{[\code{function}]\cr
Function to be interpreted as probability density function.}

\item{n}{[\code{integer(1)}]\cr
Amount of random numbers to be generated. Default is 1.}

\item{min}{[\code{numeric(1)}]\cr}

\item{max}{[\code{numeric(1)}]\cr
Interval the generated random numbers are from. Default is (0, 1).}

\item{g}{[\code{function}]\cr
Probability density function of the instrumental distribution.}

\item{rg}{[\code{function}]\cr
Function that generates random numbers from the instrumental distribution.}

\item{g.factor}{[\code{numeric(1)}]\cr
Factor for \code{g}. Default is 1.}
}
\value{
\code{n} Random numbers from the function \code{f} interpreted as probality density function on the interval (\code{min}, \code{max}).
}
\description{
Generate random numbers by rejection sampling
}
\details{
If no instrumental distribution is specified, the uniform distribution is used.

The call \code{g}(x) should return the probability of the realisation x for the instrumental distribution \code{g}.

The choice of \code{g.factor} should guarantee \code{f}(x) < \code{g.factor} * \code{g}(x).
}
\examples{
test <- function(x) x^2
rejSamp(f = test, n = 10, min = -1, max = 1)
}

